Finished test for Review model

+++ database/migrations/2020_02_21_154713_create_reviews_table.php
@@ -20,7 +20,7 @@ class CreateReviewsTable extends Migration
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('restaurant_id');            
             $table->string('comment');
-            $table->float('score');
+            $table->integer('score');
             
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');            
             $table->foreign('restaurant_id')->references('id')->on('restaurants')->onDelete('cascade');
